put HTML stuff together simplify API move hard coded parameters to default parameters
