+++ flair/visual/manifold.py
@@ -1,12 +1,6 @@
-import html
-from typing import Union, List
-
-from sklearn.manifold import TSNE
-import tqdm
 import numpy
-
-from flair.data import Sentence
-from flair.visual.html_templates import TAGGED_ENTITY, HTML_PAGE
+import tqdm
+from sklearn.manifold import TSNE
 
 
 class _Transform:
@@ -22,21 +16,6 @@ class tSNE(_Transform):
         super().__init__()
 
         self.transform = TSNE(n_components=2, verbose=1, perplexity=40, n_iter=300)
-
-
-def split_to_spans(s: Sentence):
-    orig = s.to_original_text()
-    last_idx = 0
-    spans = []
-    tagged_ents = s.get_spans('ner')
-    for ent in tagged_ents:
-        if last_idx != ent.start_pos:
-            spans.append((orig[last_idx:ent.start_pos], None))
-        spans.append((orig[ent.start_pos:ent.end_pos], ent.tag))
-        last_idx = ent.end_pos
-    if last_idx < len(orig) - 1:
-        spans.append((orig[last_idx:len(orig)], None))
-    return spans
 
 
 class Visualizer(object):
@@ -83,8 +62,8 @@ class Visualizer(object):
             for i, token in enumerate(strs):
                 prop = '<b><font color="red"> {token} </font></b>'.format(token=token)
 
-                prop = " ".join(strs[max(i - 4, 0): i]) + prop
-                prop = prop + " ".join(strs[i + 1: min(len(strs), i + 5)])
+                prop = " ".join(strs[max(i - 4, 0) : i]) + prop
+                prop = prop + " ".join(strs[i + 1 : min(len(strs), i + 5)])
 
                 contexts.append("<p>" + prop + "</p>")
 
@@ -115,9 +94,9 @@ class Visualizer(object):
                 context = '<span style="background-color: yellow"><b>{}</b></span>'.format(
                     char
                 )
-                context = "".join(sentence[max(i - 30, 0): i]) + context
+                context = "".join(sentence[max(i - 30, 0) : i]) + context
                 context = context + "".join(
-                    sentence[i + 1: min(len(sentence), i + 30)]
+                    sentence[i + 1 : min(len(sentence), i + 30)]
                 )
 
                 contexts.append(context)
@@ -148,55 +127,3 @@ class Visualizer(object):
         mpld3.plugins.connect(fig, tooltip)
 
         mpld3.save_html(fig, file)
-
-    @staticmethod
-    def render_ner_html(sentences: Union[List[Sentence], Sentence], settings=None, wrap_page=True) -> str:
-        """
-        :param sentences: single sentence or list of sentences to convert to HTML
-        :param settings: overrides and completes default settings; includes colors and labels dictionaries
-        :param wrap_page: if True method returns result of processing sentences wrapped by &lt;html&gt; and &lt;body&gt; tags, otherwise - without these tags
-        :return: HTML as a string
-        """
-        if isinstance(sentences, Sentence):
-            sentences = [sentences]
-
-        colors = {
-            "PER": "#F7FF53",
-            "ORG": "#E8902E",
-            "LOC": "#FF40A3",
-            "MISC": "#4647EB",
-            "O": "#ddd",
-        }
-
-        if settings and "colors" in settings:
-            colors.update(settings["colors"])
-
-        labels = {
-            "PER": "PER",
-            "ORG": "ORG",
-            "LOC": "LOC",
-            "MISC": "MISC",
-            "O": "O",
-        }
-
-        if settings and "labels" in settings:
-            labels.update(settings["labels"])
-
-        tagged_html = []
-        for s in sentences:
-            spans = split_to_spans(s)
-
-            for fragment, tag in spans:
-                escaped_fragment = html.escape(fragment).replace('\n', '<br/>')
-                if tag:
-                    escaped_fragment = TAGGED_ENTITY.format(entity=escaped_fragment,
-                                                            label=labels.get(tag, "O"),
-                                                            color=colors.get(tag, "#ddd"))
-                tagged_html.append(escaped_fragment)
-
-        final_text = ''.join(tagged_html)
-
-        if wrap_page:
-            return HTML_PAGE.format(text=final_text)
-        else:
-            return final_text
